[Edtor] Fixed ImGui asserts while first starting in debug mode.

+++ Runtime/Core/Settings.cpp
@@ -167,7 +167,7 @@ namespace Spartan
         Renderer* renderer = m_context->GetSubsystem<Renderer>().get();
 
         m_context->GetSubsystem<Timer>()->SetTargetFps(m_fps_limit);
-        renderer->SetOptionValue(Option_Value_Anisotropy, m_anisotropy, 0, 16);
+        renderer->SetOptionValue(Option_Value_Anisotropy, static_cast<float>(m_anisotropy), 0, 16);
         renderer->SetShadowResolution(m_shadow_map_resolution);
         renderer->SetOptions(m_renderer_flags);
     }
